Some performance tuning
We can get better performance by only updating memory addresses that
changed during the last step. For ease of use, we just update any
address that is used by an operation, or that is written to by the
command line.

We removed the sig_winch handler, and process that in normal flow. There
was some issue with things not updating correctly.  This was likely due
to not calling refresh after init_windows, but the code chnage is still
a good idea.  We delete windows on resize, then make the new windows.
Not sure if this was a memory leak or not. It is hard to tell with
ncurses.