displayoutput was next bottleneck
This should have been expected, especially for long windows. This is a
linked link and it was being reprinted every time the window changes.
That meant it was reprinted after each instruction.  That's a lot of
time printing a linked list that has not changed.  I updated the code to
only print when the contents of the list changed and when the program
started.  Moving this out of the update display loop keeps that loop
smaller and also removes a lot of looping through this list.

I renamed and moved 39numbers.sml because it really only sorts 38
numbers.  Can't believe I didn't notice that before. I must have been
tired when I named it.