Moving related data into structures
The io buffers will both be the same type of structure, instead of
one being a linked list and the other being an array.  There will
be functions to handle manipulating the structures. This will pull
the logic out of the other code.

Probably will go and create a structure for the input strings.
That way the index and the string are conceptually related to each
other.